refactor: modularize proxy architecture with SEAMS extension points
- Split 1194-line index.ts into focused modules (types, logging, env, transports, mcp-proxy)
- Add extension interfaces: ITransport, ILogger, IEnvironmentResolver, ITransportFactory
- Extract transport implementations with factory pattern for future protocols
- Consolidate duplicate command registration logic (DRY)
- Separate concerns: logging, environment resolution, transport management
- Maintain 100% backward compatibility via index.ts re-exports
- All TypeScript, ESLint, and Prettier validations pass
